[FR] Added translations
battle.json:
Added some translations

equiptype.json:
This file existed in the en folder, but isn't in the fr folder, so I
created it, I hope it's not a problem ?
The following lines aren't translate:
- Line 8, I was able to find approximately what a 'Carrier-Based
Aircraft' was, but I wasn't able to find out a correct translation in
french.
- Line 14, I don't know what a 'Daihatsu' is...

servers.json:
Translated the name of the servers.

ships.json:
I translated the name of the enemies ships, except the lines 47, 49 and
66 (see french file).
Is a a standard carrier the same thing as a carrier ?