[EN] Bm7 name revert
Pending a stronger argument than "because it isn't that hard" for
changing it. I'm sure I will regret this. We don't, for example, change
what we call E1Hard in events just because it "isn't that hard".

I am very much willing to change the name, but discussion is important.

The name was given because new players DO struggle with the quest, and
as far as I remember, it's a requirement for at least one important
questline.  And a few other, possibly cringeworthy, reasons like the
fact that it goes through a yasen node.

If new players don't know to call it Bm7, calling it "Counterattack
Fleet/Force" or anything else will not fix the problem. Calling it
anything else will also not stop new players from calling it hard.

Let's find a compromise.