[EN] quests.json backlog
shutsugeki = sortie
batsebyo = set sail

The "Press F10" thing should probably stay in desc when there's room because kc3 erases the quest in the panel, so people wouldn't see it when they need it.

Might change all the Fquest "have" to "prepare" later since it seems to be preferred but gotta look at old quests and see if there is room.