[EN] terms - MissionGetItems
The whip was cracked.

"Get" isn't used the same way in English as it is in Japanese.

While "items acquired" or something to that effect would be a more precise phrasing for this, it is implicit that items were acquired, so it's only necessary to put "items" here to save space.

I didn't make it "item(s)" because I don't like how (s) looks!