[EN] quests - minor changes
Some proofreading of those en.wiki quest titles and such

There are now so many Quartery Ranking Points quests that it's hard to remember which one is which. I've changed their EN `name` to hopefully make them easier to associate.

F89 `desc` shortened to fit into 2 lines in the overlay

Not testing the `desc` and rewards for now. I assume they are already good.