[EN] quests
people seem to be getting confused by the description, so let's try something different

an issue is that a ship class could be understood to mean either the hull code or the actual ship class, as in `Tone-class`, `Takao-class`, etc. An attempt was and is being made to avoid this confusion.